+++ HISTORY.rst
@@ -4,6 +4,19 @@
 History
 =======
 
+0.2.0 (03302016)
+----------------
+
+* Added HDR Light Rig
+* Added rename method to MatteAOV model
+* Added rename functionality to mtoatools.mattes UI
+
+  * double click to edit name
+  * enter to finish editing
+  * escape to cancel rename
+
+
+
 0.1.0 (09122015)
 ---------------------